Load SnakeYaml in a separate DependencyClassLoader. Renamed package `yaml` -> `config`.
